Scala 2.13.3 + sbt-tpolecat, fix all code for activated linting options (#1221)

+++ monix-reactive/jvm/src/main/scala/monix/reactive/observers/buffers/DropNewBufferedSubscriber.scala
@@ -177,7 +177,7 @@ private[observers] final class DropNewBufferedSubscriber[A] private (
               if (onOverflow == null || droppedCount.get() == 0)
                 null.asInstanceOf[A]
               else
-                onOverflow(droppedCount.getAndSet(0)).value() match {
+                onOverflow(droppedCount.getAndSet(0).toLong).value() match {
                   case Some(value) => value
                   case None => null.asInstanceOf[A]
                 }
