+++ monix-reactive/shared/src/test/scala/monix/reactive/internal/operators/CombineLatest5Suite.scala
@@ -31,7 +31,7 @@ object CombineLatest5Suite extends BaseOperatorSuite {
     val o2 = Observable.now(2)
     val o3 = Observable.now(3)
     val o4 = Observable.now(4)
-    val o5 = Observable.range(0, sourceCount)
+    val o5 = Observable.range(0L, sourceCount.toLong)
     val o = Observable.combineLatestMap5(o1, o2, o3, o4, o5)(_ + _ + _ + _ + _)
 
     Sample(o, count(sourceCount), sum(sourceCount), waitFirst, waitNext)
@@ -47,7 +47,7 @@ object CombineLatest5Suite extends BaseOperatorSuite {
     val o2 = Observable.now(2)
     val o3 = Observable.now(3)
     val o4 = Observable.now(4)
-    val flawed = createObservableEndingInError(Observable.range(0, sourceCount), ex)
+    val flawed = createObservableEndingInError(Observable.range(0L, sourceCount.toLong), ex)
     val o = Observable.combineLatestMap5(o1, o2, o3, o4, flawed)(_ + _ + _ + _ + _)
 
     Sample(o, count(sourceCount - 1), sum(sourceCount - 1), waitFirst, waitNext)
@@ -59,7 +59,7 @@ object CombineLatest5Suite extends BaseOperatorSuite {
     val o2 = Observable.now(2)
     val o3 = Observable.now(3)
     val o4 = Observable.now(4)
-    val o5 = Observable.range(0, sourceCount)
+    val o5 = Observable.range(0L, sourceCount.toLong)
 
     val o = Observable.combineLatestMap5(o1, o2, o3, o4, o5) { (a1, a2, a3, a4, a5) =>
       if (a5 == sourceCount - 1) throw dummy else a1 + a2 + a3 + a4 + a5
