+++ monix-reactive/shared/src/test/scala/monix/reactive/internal/operators/IntersperseSuite.scala
@@ -25,16 +25,16 @@ import scala.concurrent.duration._
 
 object IntersperseSuite extends BaseOperatorSuite {
   def createObservable(sourceCount: Int) = Option {
-    val sample = Observable.range(0, sourceCount).intersperse(1L)
-    Sample(sample, count(sourceCount), sum(sourceCount), 0.seconds, 0.seconds)
+    val sample = Observable.range(0L, sourceCount.toLong).intersperse(1L)
+    Sample(sample, count(sourceCount), sum(sourceCount).toLong, 0.seconds, 0.seconds)
   }
 
   def count(sourceCount: Int) = sourceCount * 2 - 1
   def sum(sourceCount: Int) = (sourceCount * (sourceCount - 1)) / 2 + sourceCount - 1
 
   def observableInError(sourceCount: Int, ex: Throwable) = Some {
-    val sample = createObservableEndingInError(Observable.range(0, sourceCount), ex).intersperse(0L)
-    Sample(sample, count(sourceCount), sum(sourceCount), 0.seconds, 0.seconds)
+    val sample = createObservableEndingInError(Observable.range(0L, sourceCount.toLong), ex).intersperse(0L)
+    Sample(sample, count(sourceCount), sum(sourceCount).toLong, 0.seconds, 0.seconds)
   }
 
   override def cancelableObservables(): Seq[Sample] = {
